last commit

+++ ProiectDAW/Controllers/BooksController.cs
@@ -45,13 +45,14 @@ namespace ProiectDAW.Controllers
         // PUT: api/Books/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutBook(int id, Book book)
+        public async Task<IActionResult> PutBook(int id, Book book, int adminId)
         {
-            if (id != book.Id)
+            if (id != book.Id && AdminUserExists(adminId))
             {
                 return BadRequest();
             }
 
+            book.BookType = _context.BookTypes.Find(book.BookTypeId);
             _context.Entry(book).State = EntityState.Modified;
 
             try
@@ -76,33 +77,50 @@ namespace ProiectDAW.Controllers
         // POST: api/Books
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        public async Task<ActionResult<Book>> PostBook(Book book)
+        public async Task<ActionResult<Book>> PostBook(Book book, int adminId)
         {
-            _context.Books.Add(book);
-            await _context.SaveChangesAsync();
+            if (AdminUserExists(adminId))
+            {
+                book.BookType = _context.BookTypes.Find(book.BookTypeId);
+                _context.Books.Add(book);
+                await _context.SaveChangesAsync();
 
-            return CreatedAtAction("GetBook", new { id = book.Id }, book);
+                return CreatedAtAction("GetBook", new { id = book.Id }, book);
+            }
+
+            return BadRequest();
         }
 
         // DELETE: api/Books/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteBook(int id)
+        public async Task<IActionResult> DeleteBook(int id, int adminId)
         {
-            var book = await _context.Books.FindAsync(id);
-            if (book == null)
+            if (AdminUserExists(adminId))
             {
-                return NotFound();
+                var book = await _context.Books.FindAsync(id);
+                if (book == null)
+                {
+                    return NotFound();
+                }
+
+                _context.Books.Remove(book);
+                await _context.SaveChangesAsync();
+
+                return NoContent();
             }
 
-            _context.Books.Remove(book);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
+            return BadRequest();
         }
 
         private bool BookExists(int id)
         {
             return _context.Books.Any(e => e.Id == id);
         }
+
+        private bool AdminUserExists(int id)
+        {
+            return _context.AdminUsers.Any(e => e.Id == id);
+        }
+
     }
 }
